+++ R/check_data.R
@@ -81,17 +81,6 @@ check_data_frame <- function(x, allow_missing = FALSE) {
 }
 
 
-# todo: move to utils.nest once request is implemented
-all_true <- function(lst, fcn) {
-  all(vapply(lst, fcn, TRUE))
-}
-is.logical.vector_modif <- function(x, min_size = 1) {
-  !is.null(x) &&
-    is.atomic(x) &&
-    length(x) >= min_size &&
-    all_true(x, utils.nest::is.logical.single)
-}
-
 # checks col_by and col_N to be consistent
 # col_by must be a matrix of booleans
 # checks that there are no empty levels and at least a specified number of levels
@@ -100,7 +89,7 @@ check_col_by <- function(x,
                          col_N, # nolint
                          min_num_levels = 2) {
   stopifnot(is.data.frame(col_by))
-  stopifnot(is.numeric.vector(col_N))
+  stopifnot(is_numeric_vector(col_N))
 
   if (is.data.frame(x)) {
     stopifnot(nrow(col_by) == nrow(x))
@@ -111,8 +100,8 @@ check_col_by <- function(x,
   stopifnot(
     ncol(col_by) >= min_num_levels,
     length(col_N) == ncol(col_by),
-    all(vapply(col_by, function(col) is.logical.vector_modif(col, min_size = 0), logical(1))),
-    !any(is.na(col_by)) && !("" %in% colnames(col_by))
+    all_true(col_by, is_logical_vector),
+    !anyNA(col_by) && !("" %in% colnames(col_by))
   )
 
   invisible(NULL)