+++ R/t_forest_rsp.R
@@ -92,8 +92,6 @@
 #'   table_tree = TRUE
 #' )
 #' summary(tbls)
-#'
-#' # todo: remove as_factor_keep_attributes above
 t_forest_rsp <- function(rsp,
                          col_by,
                          row_by_list = NULL,
@@ -103,7 +101,7 @@ t_forest_rsp <- function(rsp,
 
   stopifnot(
     is.logical(rsp),
-    is.null(total) || is.character.single(total),
+    is.null(total) || is_character_single(total),
     is.list(row_by_list)
   )
   do.call(check_same_n, c(list(rsp = rsp, col_by = col_by), row_by_list))
@@ -176,9 +174,6 @@ t_forest_rsp <- function(rsp,
 #' )
 #'
 t_el_forest_rsp <- function(rsp, col_by, row_name = "", dense_header = FALSE) {
-
-  # todo: we can possibly use by_hierarchical here
-
   # currently only works for factor
   col_by <- col_by_to_factor(col_by)